backend/ocr-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-service"
version = "1.0.0"
description = "Production-ready OCR service for extracting text from PDFs and images"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "OCR Service Team"}
]
keywords = ["ocr", "pdf", "text-extraction", "fastapi", "tesseract"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "python-multipart>=0.0.12",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    "pdf2image>=1.17.0",
    "PyMuPDF>=1.24.3",
    "numpy>=1.26.4",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
]
cache = [
    "redis>=5.2.0",
]
monitoring = [
    "prometheus-client>=0.21.1",
]

[project.urls]
Homepage = "https://github.com/your-org/ocr-service"
Documentation = "https://github.com/your-org/ocr-service#readme"
Repository = "https://github.com/your-org/ocr-service.git"
Issues = "https://github.com/your-org/ocr-service/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
